refactor(frontend): extract useComposerSubmission hook to satisfy code health quality gate
